adds database config in models
