+++ app/__init__.py
@@ -1,11 +1,24 @@
 from flask import Flask
+from flask_sqlalchemy import SQLAlchemy
+from flask_migrate import Migrate
 
-
+db = SQLAlchemy
+migrate = Migrate()
+DATABASE = 'postgresql+psycopg2://postgres:postgres@localhost:5432/solar_system_development'
 
 def create_app(test_config=None):
     app = Flask(__name__)
 
+    app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
+    app.config['SQLALCHEMY_DATABSE_URI'] = DATABASE
+
+    db.init_app(app)
+    migrate.init_app(app,db)
+
     from .routes import planets_bp
     app.register_blueprint(planets_bp)
 
+    from app.models.planet import Planet
+
     return app
+
